moved auth.css to css folder
